requirements.txt
pytube3==9.6.4
typing-extensions==3.7.4.3
